requirements.txt
lightning[pytorch-extra]==2.1
wandb==0.17.2
transformers==4.41.2
pandas==2.2.2
numpy==1.24.3
tokenizers==0.19.1
monai==1.3.1
numba==0.59.1
matplotlib==3.9.0
ffcv==1.0.2
notebook==7.2.1
pysimdjson==6.0.2
ipywidgets==8.1.3
tables==3.9.2
scikit-learn==1.13.1
seaborn==0.13.2